Move the modules to be co-located where used ...
While the data schema generation may be fairly general it is only used in the
data inpsection module. Attempting to co-locate modules to more
appropriate hub locations to avoid having utils be polluted namespace
